[refactor] 공통 reducer 분리
- 기능 유형에 동일하게 사용될 수 있는 reducer 생성 함수 분리
- 공통으로 사용되는 Feature 상태 타입도 Common으로 분리
