+++ src/hooks/sidebar/useFeatureTypeItem.ts
@@ -1,11 +1,11 @@
 import { useSelector } from 'react-redux';
 import { RootState } from '../../store';
-import { PoiType } from '../../store/style/poiReducer';
+import { FeatureState } from '../../store/common/type';
 import { FeatureNameType } from '../../utils/rendering-data/featureTypeData';
 
 interface useFeatureTypeItemType {
   // 나중에 | 연산으로 다양한 타입으로 수정 필요
-  featureList: PoiType;
+  featureList: FeatureState;
 }
 
 export interface useFeatureTypeItemProps {
@@ -17,7 +17,7 @@ function useFeatureTypeItem({
 }: useFeatureTypeItemProps): useFeatureTypeItemType {
   const featureList = useSelector<RootState>(
     (state) => state[featureName]
-  ) as PoiType; // 나중에 다양한 타입으로 수정 필요
+  ) as FeatureState; // 나중에 다양한 타입으로 수정 필요
 
   return {
     featureList,
