+++ src/store/common/type.ts
@@ -1,0 +1,9 @@
+import { LabelType, CommonType } from './commonProperties';
+
+export interface FeatureState {
+  [name: string]: {
+    isChanged: boolean;
+    section: CommonType;
+    label: LabelType;
+  };
+}
